Implemented Web3 wallet, while fixing base classes
