+++ python/src/wallets/goat-wallet-evm/goat_wallets/evm/evm_wallet_client.py
@@ -2,7 +2,7 @@ from abc import abstractmethod
 from typing import Dict
 
 from goat.types.chain import EvmChain
-from goat.classes.wallet_client_base import WalletClientBase
+from goat.classes.wallet_client_base import Signature, WalletClientBase
 
 from .types import EVMTransaction, EVMReadRequest, EVMReadResult, EVMTypedData
 
@@ -14,21 +14,21 @@ class EVMWalletClient(WalletClientBase):
         pass
 
     @abstractmethod
-    async def send_transaction(self, transaction: EVMTransaction) -> Dict[str, str]:
+    def send_transaction(self, transaction: EVMTransaction) -> Dict[str, str]:
         """Send a transaction on the EVM chain."""
         pass
 
     @abstractmethod
-    async def read(self, request: EVMReadRequest) -> EVMReadResult:
+    def read(self, request: EVMReadRequest) -> EVMReadResult:
         """Read data from a smart contract."""
         pass
 
     @abstractmethod
-    async def resolve_address(self, address: str) -> str:
+    def resolve_address(self, address: str) -> str:
         """Resolve an address to its canonical form."""
         pass
 
     @abstractmethod
-    async def sign_typed_data(self, data: EVMTypedData) -> str:
+    def sign_typed_data(self, data: EVMTypedData) -> Signature:
         """Sign typed data according to EIP-712."""
         pass
